requirements.txt
python-dotenv
huggingface_hub
mlx-lm